cleanup

+++ spvm_api.c
@@ -638,12 +638,12 @@ void SPVM_API_call_sub(SPVM* spvm, SPVM_ENV* env, const char* sub_abs_name) {
         goto *jump[*pc];
       case_SPVM_BYTECODE_C_CODE_PUSH_SHORT:
         operand_stack_top++;
-        call_stack[operand_stack_top].short_value = (int16_t)((int16_t)(*(pc + 1) << 8) +  *(pc + 2));
+        call_stack[operand_stack_top].short_value = (int16_t)((*(pc + 1) << 8) +  *(pc + 2));
         pc += 3;
         goto *jump[*pc];
       case_SPVM_BYTECODE_C_CODE_PUSH_BYTE_TO_INT:
         operand_stack_top++;
-        call_stack[operand_stack_top].int_value = (int8_t)*(pc + 1);
+        call_stack[operand_stack_top].int_value = (int32_t)*(pc + 1);
         pc += 2;
         goto *jump[*pc];
       case_SPVM_BYTECODE_C_CODE_PUSH_SHORT_TO_LONG:
@@ -658,7 +658,7 @@ void SPVM_API_call_sub(SPVM* spvm, SPVM_ENV* env, const char* sub_abs_name) {
         goto *jump[*pc];
       case_SPVM_BYTECODE_C_CODE_PUSH_SHORT_TO_INT:
         operand_stack_top++;
-        call_stack[operand_stack_top].int_value = (int16_t)((int16_t)(*(pc + 1) << 8) +  *(pc + 2));
+        call_stack[operand_stack_top].int_value = (int32_t)((int16_t)(*(pc + 1) << 8) +  *(pc + 2));
         pc += 3;
         goto *jump[*pc];
       case_SPVM_BYTECODE_C_CODE_LOAD_CONSTANT:
